+++ NEWS.md
@@ -12,4 +12,4 @@
 
 * Added `weights` input to `orsf`, allowing users to over or under fit `orsf` to specific data in their training set.
 
-* Added `chf` and `mort` options to `predict.orsf_fit()`
+* Added `chf` and `mort` options to `predict.orsf_fit()`. Mortality predictions are not fully implemented yet - they are not supported in partial dependence or out-of-bag error estimates. These features will be added in a future update.